twic

Module value

The Twic Value enum, representing all possible Twic values.

Enums

Number

Represents a Twic number, which can be an integer, float (excluding NaN and Infinity), nan, and [+/-]inf.

Value

Represents a Twic value.

ValueIndexError

Errors that can occur when indexing into a Value.

Traits

IndexInto

Trait for types that can index into a Value.

Type Aliases

IndexMutResult

Mutable result type for indexing into a Value.

IndexResult

Result type for indexing into a Value.

Map

A map value in Twic, mapping string keys to Values.