twic::value

Trait IndexInto

pub trait IndexInto {
    // Required methods
    fn index_into<'a>(&self, value: &'a Value) -> IndexResult<'a>;
    fn index_into_mut<'a>(&self, value: &'a mut Value) -> IndexMutResult<'a>;
    fn index_into_or_insert<'a>(
        &self,
        value: &'a mut Value,
    ) -> IndexMutResult<'a>;
}

Trait for types that can index into a Value.

Required Methods

fn index_into<'a>(&self, value: &'a Value) -> IndexResult<'a>

Indexes into the given Value, returning a reference to the indexed Value or an error if the index is invalid or the Value is not indexable.

fn index_into_mut<'a>(&self, value: &'a mut Value) -> IndexMutResult<'a>

Indexes into the given mutable Value, returning a mutable reference to the indexed Value or an error if the index is invalid or the Value is not indexable.

fn index_into_or_insert<'a>(&self, value: &'a mut Value) -> IndexMutResult<'a>

Indexes into the given mutable Value, inserting a new Value if the index does not exist. Returns a mutable reference to the indexed Value, or an error if the index is invalid or the Value is not indexable.